+++ R/platjam-import.R
@@ -505,6 +505,7 @@ nmatlist2heatmaps <- function
  column_title_gp=grid::gpar(fontsize=12),
  lens=-2,
  anno_lens=8,
+ pos_line=FALSE,
  seed=123,
  ht_gap=grid::unit(3, "mm"),
  profile_value=c("mean", "sum", "abs_mean", "abs_sum"),
@@ -648,6 +649,13 @@ nmatlist2heatmaps <- function
          "str(transform):");
       print(str(transform));
    }
+
+   ## pos_line
+   if (length(pos_line) == 0) {
+      pos_line <- FALSE;
+   }
+   pos_line <- rep(pos_line,
+      length.out=length(nmatlist));
 
    ## optional signal_ceiling
    if (length(signal_ceiling) > 0) {
@@ -1474,7 +1482,7 @@ nmatlist2heatmaps <- function
 
       EH <- EnrichedHeatmap::EnrichedHeatmap(imat[rows,],
          split=partition[rows],
-         pos_line=FALSE,
+         pos_line=pos_line[[i]],
          use_raster=use_raster,
          col=colramp,
          border=border[[i]],